refactor: reduce logging noise
The log messages about the ignored committed offsets are not really useful
unless we are debugging problems with that. Therefore, we could simply
leave an error log message for when the offset committing goes wrong, and
a debug one when we really want to know that offsets are being committed.

RHCLOUD-36895