refactor: increment errors counter when retries have been depleted
Right now if an error occurs when attempting to store the payload in the
first attempt, the "processing errors" counter is incremented, even if
that payload might be properly stored in a second attempt.

This might trigger alerts that actually do not really need any
attention, since the payloads might be getting updated correctly.

RHCLOUD-36896